needed cast for my machine for some reason

+++ ls/_ls.c
@@ -58,7 +58,7 @@ void long_print(char *path)
 	}
 
 	printf("%s %2lu %-8s %-8s %5ld %s %s\n",
-		   perms, buf.st_nlink, uname, gname, buf.st_size, time_str, file_name);
+		   perms, (unsigned long)buf.st_nlink, uname, gname, buf.st_size, time_str, file_name);
 }
 
 /**
